🔥 move legacy cost function to tests and remove some of the legacy tests
