+++ test/test_zx.cpp
@@ -13,7 +13,6 @@
 #include "qasm3/Importer.hpp"
 #include "zx/ZXDefinitions.hpp"
 
-#include <algorithm>
 #include <gtest/gtest.h>
 #include <iostream>
 #include <memory>
@@ -213,8 +212,6 @@ TEST_F(ZXTest, ZXWrongAncilla) {
 }
 
 TEST_F(ZXTest, ZXConfiguredForInvalidCircuitParallel) {
-  using namespace qc::literals;
-
   auto qc = qc::QuantumComputation(4);
   qc.mcx({1, 2, 3}, 0);
 
@@ -226,8 +223,6 @@ TEST_F(ZXTest, ZXConfiguredForInvalidCircuitParallel) {
 }
 
 TEST_F(ZXTest, ZXConfiguredForInvalidCircuitSequential) {
-  using namespace qc::literals;
-
   auto qc = qc::QuantumComputation(4);
   qc.mcx({1, 2, 3}, 0);
 
@@ -251,49 +246,6 @@ TEST_F(ZXTest, GlobalPhase) {
 
   EXPECT_EQ(ecm->getResults().equivalence,
             ec::EquivalenceCriterion::EquivalentUpToGlobalPhase);
-}
-
-class ZXTestCompFlow : public testing::TestWithParam<std::string> {
-protected:
-  qc::QuantumComputation qcOriginal;
-  qc::QuantumComputation qcTranspiled;
-  ec::Configuration config{};
-
-  std::unique_ptr<ec::EquivalenceCheckingManager> ecm;
-
-  std::string testOriginalDir = "./circuits/original/";
-  std::string testTranspiledDir = "./circuits/transpiled/";
-
-  void SetUp() override {
-    config.execution.parallel = false;
-    config.execution.runConstructionChecker = false;
-    config.execution.runAlternatingChecker = false;
-    config.execution.runSimulationChecker = false;
-    config.execution.runZXChecker = true;
-
-    qcOriginal =
-        qasm3::Importer::importf(testOriginalDir + GetParam() + ".qasm");
-    qcTranspiled = qasm3::Importer::importf(testTranspiledDir + GetParam() +
-                                            "_transpiled.qasm");
-  }
-};
-
-INSTANTIATE_TEST_SUITE_P(
-    ZXTestCompFlow, ZXTestCompFlow,
-    testing::Values("c2_181", "rd73_312", "sym9_317", "mod5adder_306",
-                    "rd84_313"),
-    [](const testing::TestParamInfo<ZXTestCompFlow::ParamType>& inf) {
-      auto s = inf.param;
-      std::replace(s.begin(), s.end(), '-', '_');
-      return s;
-    });
-
-TEST_P(ZXTestCompFlow, EquivalenceCompilationFlow) {
-  ecm = std::make_unique<ec::EquivalenceCheckingManager>(qcOriginal,
-                                                         qcTranspiled, config);
-  ecm->run();
-  std::cout << ecm->getResults() << "\n";
-  EXPECT_TRUE(ecm->getResults().consideredEquivalent());
 }
 
 TEST(ZXTestsMisc, IdentityNotHadamard) {
